+++ processor/spanmetricsprocessor/processor_test.go
@@ -343,7 +343,7 @@ func verifyConsumeMetricsInput(input pdata.Metrics, t *testing.T) bool {
 		assert.Equal(t, "calls_total", m.At(mi).Name())
 
 		data := m.At(mi).Sum()
-		assert.Equal(t, pdata.AggregationTemporalityCumulative, data.AggregationTemporality())
+		assert.Equal(t, pdata.MetricAggregationTemporalityCumulative, data.AggregationTemporality())
 		assert.True(t, data.IsMonotonic())
 
 		dps := data.DataPoints()
@@ -363,7 +363,7 @@ func verifyConsumeMetricsInput(input pdata.Metrics, t *testing.T) bool {
 		assert.Equal(t, "latency", m.At(mi).Name())
 
 		data := m.At(mi).Histogram()
-		assert.Equal(t, pdata.AggregationTemporalityCumulative, data.AggregationTemporality())
+		assert.Equal(t, pdata.MetricAggregationTemporalityCumulative, data.AggregationTemporality())
 
 		dps := data.DataPoints()
 		require.Equal(t, 1, dps.Len())